requirements-embed.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
faiss-cpu
python-multipart
